Move mount and minion classes to dedicated packages
